fix quantile crash on bfloat16 by casting to float
torch.quantile() only supports float/double. The downcast diff path
produces bfloat16 abs_diff tensors, so _quantile_or_none must .float()
before computing.

+++ python/sglang/srt/debug_utils/comparator/tensor_comparator/comparator.py
@@ -104,7 +104,7 @@ def _compute_tensor_stats(x: torch.Tensor) -> TensorStats:
 
 
 def _quantile_or_none(x: torch.Tensor, *, q: float, include: bool) -> Optional[float]:
-    return torch.quantile(x, q).item() if include else None
+    return torch.quantile(x.float(), q).item() if include else None
 
 
 def _compute_diff(
